requirements.txt
numpy
opencv-python
matplotlib
tensorflow==2.15.0
tqdm
scikit-image
ffmpeg-python
ipython